requirements.txt
amqp==5.1.1
asgiref==3.6.0
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.26.0
cron-descriptor==1.2.35
Django==4.1.7
django-celery-beat==2.5.0
django-timezone-field==5.0
djangorestframework==3.14.0
ffmpeg-python==0.2.0
filelock==3.10.0
future==0.18.3
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
lit==15.0.7
llvmlite==0.39.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper==20230314
prompt-toolkit==3.0.38
pycodestyle==2.10.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
sympy==1.11.1
tiktoken==0.3.1
tomli==2.0.1
torch==2.0.0
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
